+++ releasenotes.md
@@ -49,6 +49,7 @@
 - Calling a constant void macro inside a macro stops it from being constant. #3410
 - Using io::struct_to_format with `$force_dump = true` failed to compile.
 - `$foo += 1` would not do a copy, leading to incorrect update of `$foo`. #3400
+- `foreach (foo::Type t : x)` would not parse properly. #3423
 
 ## 0.8.2 Change list